+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v2.0.37
+* Added `ButtonType.search` and `ButtonType.searchAlt` to CustomButton with a search icon and outlined secondary style
+* Allowed `type="submit"` to be passed to CustomButton
+* Added story for Search CustomButton
+
 ## v2.0.36
 * Added component HyperLink for links with optional icon
 * Added stories for HyperLink and FormLink